Add files via upload
Hello, Im jyomama28. I recently created my website swap it as a script kiddie lol. I worked on it and im just gonna post it to my new github account. this is in the early stages and im just a beginner. Please visit and test out my website, Also let me know if it has any bugs that need fixing or something idk.